+++ corda-testacles-mocknetwork/src/main/kotlin/com/github/manosbatsis/corda/testacles/mocknetwork/NodeHandles.kt
@@ -21,35 +21,26 @@
  */
 package com.github.manosbatsis.corda.testacles.mocknetwork
 
+import com.github.manosbatsis.corda.testacles.common.jupiter.AbstractNodesMap
 import net.corda.core.identity.CordaX500Name
 import net.corda.core.identity.Party
 import net.corda.testing.node.StartedMockNode
 
+
 /**
  * The[mockNetwork]'s [StartedMockNode]s, each mapped by multiple keys:
  *
+ * - the original input key
  * - identity ([Party])
  * - identity name ([CordaX500Name]
  * - identity name (string representation)
  * - organization name
+ * - organization name with lower case first char, without spaces
  * - organization name in lower case without spaces
  *
- * In other words `nodes["O=PartyA, L=Athens, C=GR"]` is the same as `nodes["partya"]`
+ * In other words `getNode["O=PartyA, L=Athens, C=GR"]` is the same as `getNode["partya"]`
  */
-open class NodeHandles(
-        private val nodes: Map<Any, StartedMockNode> = emptyMap()
-){
-
-    fun findNodeByIdentity(identity: CordaX500Name): StartedMockNode? =
-            nodes[identity]
-
-    fun getNodeByIdentity(identity: CordaX500Name): StartedMockNode =
-            findNodeByIdentity(identity) ?: error("No matching node found for identity: $identity")
-
-    fun findNodeByKey(key: String): StartedMockNode? = nodes[key]
-
-    fun getNodeByKey(key: String): StartedMockNode =
-            findNodeByKey(key) ?: error("No matching node found for key: $key")
-
-
+class NodeHandles(input: Map<String, StartedMockNode>): AbstractNodesMap<StartedMockNode>(input){
+    override fun getIdentity(of: StartedMockNode): Party =
+            of.info.legalIdentities.first()
 }